Fibonacci in Java
